+++ packages/amplify-graphql-relational-transformer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.5](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-relational-transformer@2.3.4...@aws-amplify/graphql-relational-transformer@2.3.5) (2024-01-30)
+
+**Note:** Version bump only for package @aws-amplify/graphql-relational-transformer
+
+## [2.3.4](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-relational-transformer@2.3.3...@aws-amplify/graphql-relational-transformer@2.3.4) (2024-01-22)
+
+**Note:** Version bump only for package @aws-amplify/graphql-relational-transformer
+
 ## [2.3.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-relational-transformer@2.3.2...@aws-amplify/graphql-relational-transformer@2.3.3) (2023-12-21)
 
 ### Bug Fixes