+++ packages/amplify-graphql-transformer-test-utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.4](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-test-utils@0.4.3...@aws-amplify/graphql-transformer-test-utils@0.4.4) (2024-01-30)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-test-utils
+
+## [0.4.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-test-utils@0.4.2...@aws-amplify/graphql-transformer-test-utils@0.4.3) (2024-01-22)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-test-utils
+
 ## [0.4.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-test-utils@0.4.1...@aws-amplify/graphql-transformer-test-utils@0.4.2) (2023-12-18)
 
 **Note:** Version bump only for package @aws-amplify/graphql-transformer-test-utils